refact: move metrics to module
